feat : add ROS2 Message Classes

+++ src/main/kotlin/net/wavem/uvc/ros2/geometry_msgs/msg/Point.kt
@@ -1,8 +1,8 @@
 package net.wavem.uvc.ros2.geometry_msgs.msg
 
 data class Point(
-    val x: Float,
-    val y: Float,
-    val z: Float
+    val x: Double,
+    val y: Double,
+    val z: Double
 ) {
 }
